requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
python-matlab-engine==3.11.0
psutil==5.9.6